# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/subscription_plans:
    get:
      tags:
        - SubscriptionPlans
      summary: List SubscriptionPlans
      description: Get a paginated list of SubscriptionPlans with optional filtering and sorting
      operationId: listSubscriptionPlans
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionPlanList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SubscriptionPlans
      summary: Create SubscriptionPlan
      description: Create a new SubscriptionPlan
      operationId: createSubscriptionPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionPlanRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionPlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/subscription_plan_lines:
    get:
      tags:
        - SubscriptionPlanLines
      summary: List SubscriptionPlanLines
      description: Get a paginated list of SubscriptionPlanLines with optional filtering and sorting
      operationId: listSubscriptionPlanLines
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionPlanLineList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SubscriptionPlanLines
      summary: Create SubscriptionPlanLine
      description: Create a new SubscriptionPlanLine
      operationId: createSubscriptionPlanLine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionPlanLineRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionPlanLine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    SubscriptionPlan:
      type: object
      required:
        - id
        - plan_code
        - name
        - billing_cycle
        - billing_day
        - currency
        - receivable_account_id
        - price
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        plan_code:
          type: string
          maxLength: 40
        name:
          type: string
        description:
          nullable: true
          type: string
        billing_cycle:
          type: string
        billing_day:
          type: integer
          format: int32
        currency:
          type: string
          maxLength: 3
        receivable_account_id:
          type: string
          format: uuid
        price:
          type: string
        trial_days:
          nullable: true
          type: integer
          format: int32
        status:
          type: string
        metadata:
          type: object
...
    CreateSubscriptionPlanRequest:
      type: object
      required:
        - plan_code
        - name
        - billing_cycle
        - billing_day
        - currency
        - receivable_account_id
        - price
        - status
        - metadata
      properties:
        plan_code:
          type: string
          maxLength: 40
        name:
          type: string
        description:
          nullable: true
          type: string
        billing_cycle:
          type: string
        billing_day:
          type: integer
          format: int32
        currency:
          type: string
          maxLength: 3
        receivable_account_id:
          type: string
          format: uuid
        price:
          type: string
        trial_days:
          nullable: true
          type: integer
          format: int32
        status:
          type: string
        metadata:
          type: object
...
    SubscriptionPlanList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionPlan'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    SubscriptionPlanLine:
      type: object
      required:
        - id
        - plan_id
        - item_id
        - account_id
        - quantity
        - unit_price
        - metadata
      properties:
        id:
          type: string
          format: uuid
        plan_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
        quantity:
          type: string
        unit_price:
          type: string
        metadata:
          type: object
...
    CreateSubscriptionPlanLineRequest:
      type: object
      required:
        - plan_id
        - item_id
        - account_id
        - quantity
        - unit_price
        - metadata
      properties:
        plan_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
        quantity:
          type: string
        unit_price:
          type: string
        metadata:
          type: object
...
    SubscriptionPlanLineList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionPlanLine'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []